requirements.txt
fastapi
uvicorn
pydantic
reportlab
qrcode
boto3
sqlalchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
python-multipart